Separating out operator unit tests.
